requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile --output-file=src/requirements.txt src/requirements/requirements.in
#
asgiref==3.8.1
    # via
    #   django
    #   django-allauth
attrs==24.2.0
    # via
    #   jsonschema
    #   referencing
brotli==1.1.0
    # via fonttools
build==1.2.2.post1
    # via pip-tools
certifi==2024.8.30
    # via requests
cffi==1.17.1
    # via weasyprint
charset-normalizer==3.4.0
    # via requests
click==8.1.7
    # via pip-tools
cssselect2==0.7.0
    # via weasyprint
dj-database-url==2.3.0
    # via -r src/requirements/requirements.in
dj-notification-api==1.2.0
    # via -r src/requirements/requirements.in
django==5.1.3
    # via
    #   -r src/requirements/requirements.in
    #   dj-database-url
    #   dj-notification-api
    #   django-allauth
    #   django-filter
    #   django-pwa
    #   django-storages
    #   djangorestframework
    #   drf-spectacular
    #   drf-spectacular-sidecar
    #   slippers
django-admin-interface==0.29.1
    # via -r src/requirements/requirements.in
django-allauth==65.2.0
    # via -r src/requirements/requirements.in
django-allauth-ui==1.5.1
    # via -r src/requirements/requirements.in
django-colorfield==0.11.0
    # via django-admin-interface
django-filter==24.3
    # via
    #   -r src/requirements/requirements.in
    #   dj-notification-api
django-pwa==2.0.1
    # via -r src/requirements/requirements.in
django-storages==1.14.4
    # via -r src/requirements/requirements.in
django-widget-tweaks==1.5.0
    # via django-allauth-ui
djangorestframework==3.15.2
    # via
    #   -r src/requirements/requirements.in
    #   dj-notification-api
    #   drf-extensions
    #   drf-spectacular
drf-extensions==0.7.1
    # via -r src/requirements/requirements.in
drf-spectacular==0.27.2
    # via -r src/requirements/requirements.in
drf-spectacular-sidecar==2024.11.1
    # via -r src/requirements/requirements.in
fonttools[woff]==4.55.0
    # via weasyprint
gunicorn==23.0.0
    # via -r src/requirements/requirements.in
html5lib==1.1
    # via weasyprint
idna==3.10
    # via requests
inflection==0.5.1
    # via drf-spectacular
jsonschema==4.23.0
    # via drf-spectacular
jsonschema-specifications==2024.10.1
    # via jsonschema
packaging==24.2
    # via
    #   build
    #   gunicorn
pillow==11.0.0
    # via
    #   -r src/requirements/requirements.in
    #   django-colorfield
    #   weasyprint
pip-tools==7.4.1
    # via -r src/requirements/requirements.in
psycopg2-binary==2.9.10
    # via -r src/requirements/requirements.in
pycparser==2.22
    # via cffi
pydyf==0.11.0
    # via weasyprint
pyphen==0.17.0
    # via weasyprint
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
python-decouple==3.8
    # via -r src/requirements/requirements.in
python-dotenv==1.0.1
    # via -r src/requirements/requirements.in
python-slugify==8.0.4
    # via django-admin-interface
pyyaml==6.0.2
    # via
    #   drf-spectacular
    #   slippers
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.3
    # via -r src/requirements/requirements.in
rpds-py==0.21.0
    # via
    #   jsonschema
    #   referencing
six==1.16.0
    # via html5lib
slippers==0.6.2
    # via django-allauth-ui
sqlparse==0.5.2
    # via django
text-unidecode==1.3
    # via python-slugify
tinycss2==1.4.0
    # via
    #   cssselect2
    #   weasyprint
typeguard==2.13.3
    # via slippers
typing-extensions==4.12.2
    # via
    #   dj-database-url
    #   slippers
uritemplate==4.1.1
    # via drf-spectacular
urllib3==2.2.3
    # via requests
weasyprint==62.3
    # via -r src/requirements/requirements.in
webencodings==0.5.1
    # via
    #   cssselect2
    #   html5lib
    #   tinycss2
wheel==0.45.0
    # via pip-tools
whitenoise==6.9.0
    # via -r src/requirements/requirements.in
zopfli==0.2.3.post1
    # via fonttools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
